requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
flasgger==0.9.7.1
